extracting pure dns server to specific package
