+++ src/main/java/com/mageddo/dnsproxyserver/config/dataprovider/vo/ConfigJson.java
@@ -1,7 +1,7 @@
 package com.mageddo.dnsproxyserver.config.dataprovider.vo;
 
 import com.mageddo.dnsproxyserver.config.Config;
-import com.mageddo.dnsproxyserver.server.dns.SimpleServer;
+import com.mageddo.dnsserver.SimpleServer;
 import com.mageddo.net.IpAddr;
 
 import java.net.URI;
