+++ src/main/java/com/mageddo/dnsproxyserver/di/module/ModuleMap.java
@@ -3,7 +3,7 @@ package com.mageddo.dnsproxyserver.di.module;
 import com.mageddo.dnsproxyserver.config.dataprovider.PersistentConfigDAO;
 import com.mageddo.dnsproxyserver.docker.dataprovider.ContainerFacade;
 import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
-import com.mageddo.dnsproxyserver.server.dns.ServerStarter;
+import com.mageddo.dnsserver.ServerStarter;
 import com.mageddo.dnsproxyserver.server.dns.solver.SolverLocalDB;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.DockerDAO;
 import dagger.Binds;
