+++ src/main/java/com/mageddo/dnsserver/UDPServer.java
@@ -1,7 +1,8 @@
-package com.mageddo.dnsproxyserver.server.dns;
+package com.mageddo.dnsserver;
 
 import com.mageddo.commons.concurrent.ThreadPool;
 import com.mageddo.commons.io.IoUtils;
+import com.mageddo.dnsproxyserver.server.dns.Messages;
 import lombok.extern.slf4j.Slf4j;
 import org.xbill.DNS.Message;
