+++ src/test/java/testing/templates/ConfigTemplates.java
@@ -3,7 +3,7 @@ package testing.templates;
 import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.LogLevel;
 import com.mageddo.dnsproxyserver.config.dataprovider.vo.ConfigEnv;
-import com.mageddo.dnsproxyserver.server.dns.SimpleServer;
+import com.mageddo.dnsserver.SimpleServer;
 
 import java.nio.file.Paths;
