+++ src/test/java/testing/templates/SocketClientTemplates.java
@@ -1,7 +1,6 @@
 package testing.templates;
 
-import com.mageddo.dnsproxyserver.server.dns.SocketClient;
-import org.mockito.Mockito;
+import com.mageddo.dnsserver.SocketClient;
 
 import java.io.InputStream;
 import java.io.OutputStream;
